+++ dist-deno/libs/publicTypes.ts
@@ -6,5 +6,8 @@ export type NullableRules = {
 } | {
     ifEmptyString: null;
 };
+export type UndefinableRules = {
+    ifUndefined: undefined;
+};
 export type ErrorHandler<T = unknown> = (err: ValueSchemaError) => T | never;
 export type FinishHandler = () => void;