+++ dist-deno/schemas/boolean.ts
@@ -1,5 +1,9 @@
-import { NullableRules } from "../libs/publicTypes.ts";
+import { NullableRules, UndefinableRules } from "../libs/publicTypes.ts";
 import { BooleanSchema, RulesForBoolean } from "../schemaClasses/BooleanSchema.ts";
+/** schema for boolean or null or undefined */
+export function boolean(options: RulesForBoolean & NullableRules & UndefinableRules): BooleanSchema<null | undefined>;
+/** schema for boolean or undefined */
+export function boolean(options: RulesForBoolean & UndefinableRules): BooleanSchema<undefined>;
 /** schema for boolean or null */
 export function boolean(rules: RulesForBoolean & NullableRules): BooleanSchema<null>;
 /** schema for boolean */