+++ dist-deno/schemas/email.ts
@@ -1,5 +1,9 @@
-import { NullableRules } from "../libs/publicTypes.ts";
+import { NullableRules, UndefinableRules } from "../libs/publicTypes.ts";
 import { EmailSchema, RulesForEmail } from "../schemaClasses/EmailSchema.ts";
+/** schema for email or null or undefined */
+export function email(options: RulesForEmail & NullableRules & UndefinableRules): EmailSchema<null | undefined>;
+/** schema for email or undefined */
+export function email(options: RulesForEmail & UndefinableRules): EmailSchema<undefined>;
 /** schema for email or null */
 export function email(rules: RulesForEmail & NullableRules): EmailSchema<null>;
 /** schema for email */