+++ dist-deno/schemas/enumeration.ts
@@ -1,5 +1,9 @@
-import { NullableRules } from "../libs/publicTypes.ts";
+import { NullableRules, UndefinableRules } from "../libs/publicTypes.ts";
 import { EnumerationSchema, RulesForEnumeration } from "../schemaClasses/EnumerationSchema.ts";
+/** schema for enum-like (enum / union) or null or undefined */
+export function enumeration<E = unknown>(options: RulesForEnumeration<E> & NullableRules & UndefinableRules): EnumerationSchema<E, null | undefined>;
+/** schema for enum-like (enum / union) or undefined */
+export function enumeration<E = unknown>(options: RulesForEnumeration<E> & UndefinableRules): EnumerationSchema<E, undefined>;
 /** schema for enum-like (enum / union) or null */
 export function enumeration<E = unknown>(rules: RulesForEnumeration<E> & NullableRules): EnumerationSchema<E, null>;
 /** schema for enum-like (enum / union) */
@@ -9,6 +13,6 @@ export function enumeration<E = unknown>(rules: RulesForEnumeration<E>): Enumera
  * @param rules rules
  * @returns schema
  */
-export function enumeration<E = unknown>(rules: RulesForEnumeration<E>): EnumerationSchema<E, null> {
+export function enumeration<E = unknown>(rules: RulesForEnumeration<E>): EnumerationSchema<E> {
     return new EnumerationSchema<E>(rules);
 }