+++ dist-deno/schemas/number.ts
@@ -1,6 +1,10 @@
-import { NullableRules } from "../libs/publicTypes.ts";
+import { NullableRules, UndefinableRules } from "../libs/publicTypes.ts";
 import { NumberSchema, RulesForNumber } from "../schemaClasses/NumberSchema.ts";
 export { NUMBER } from "../schemaClasses/NumberSchema.ts";
+/** schema for number or null or undefined */
+export function number(options: RulesForNumber & NullableRules & UndefinableRules): NumberSchema<null | undefined>;
+/** schema for number or undefined */
+export function number(options: RulesForNumber & UndefinableRules): NumberSchema<undefined>;
 /** schema for number or null */
 export function number(rules: RulesForNumber & NullableRules): NumberSchema<null>;
 /** schema for number */
@@ -12,6 +16,6 @@ export function number(): NumberSchema;
  * @param rules rules
  * @returns schema
  */
-export function number(rules: RulesForNumber = {}): NumberSchema<null> {
+export function number(rules: RulesForNumber = {}): NumberSchema {
     return new NumberSchema(rules);
 }