+++ dist-deno/schemas/numericString.ts
@@ -1,6 +1,10 @@
-import { NullableRules } from "../libs/publicTypes.ts";
+import { NullableRules, UndefinableRules } from "../libs/publicTypes.ts";
 import { NumericStringSchema, RulesForNumericString } from "../schemaClasses/NumericStringSchema.ts";
 export { NUMERIC_STRING } from "../schemaClasses/NumericStringSchema.ts";
+/** schema for numeric string or null or undefined */
+export function numericString(options: RulesForNumericString & NullableRules & UndefinableRules): NumericStringSchema<null | undefined>;
+/** schema for numeric string or undefined */
+export function numericString(options: RulesForNumericString & UndefinableRules): NumericStringSchema<undefined>;
 /** schema for numeric string or null */
 export function numericString(rules: RulesForNumericString & NullableRules): NumericStringSchema<null>;
 /** schema for numeric string */