+++ dist-deno/schemas/string.ts
@@ -1,6 +1,10 @@
-import { NullableRules } from "../libs/publicTypes.ts";
+import { NullableRules, UndefinableRules } from "../libs/publicTypes.ts";
 import { RulesForString, StringSchema } from "../schemaClasses/StringSchema.ts";
 export { STRING } from "../schemaClasses/StringSchema.ts";
+/** schema for string or null or undefined */
+export function string(options: RulesForString & NullableRules & UndefinableRules): StringSchema<null | undefined>;
+/** schema for string or undefined */
+export function string(options: RulesForString & UndefinableRules): StringSchema<undefined>;
 /** schema for string or null */
 export function string(rules: RulesForString & NullableRules): StringSchema<null>;
 /** schema for string */