+++ test/schemas/boolean.test.ts
@@ -198,6 +198,14 @@ function testIfUndefined(): void
 			}).applyTo(undefined)
 		).toEqual(true);
 	});
+	it("should be undefined", () =>
+	{
+		expect(
+			vs.boolean({
+				ifUndefined: undefined,
+			}).applyTo(undefined)
+		).toBeUndefined();
+	});
 	it("should cause error(s)", () =>
 	{
 		expect(() =>